+++ packages/@webex/plugin-meetings/test/unit/spec/annotation/index.ts
@@ -144,7 +144,10 @@ describe('live-annotation', () => {
 
         annotationService.eventDataProcessor({data: {relayType: 'annotation.client', request:{value:{encryptionKeyUrl:"encryptionKeyUrl"}}}});
 
-        assert.calledOnceWithExactly(spy, {value:{encryptionKeyUrl:"encryptionKeyUrl"}});
+        assert.calledOnceWithExactly(spy, {
+          relayType: 'annotation.client',
+          request: { value: { encryptionKeyUrl: 'encryptionKeyUrl' } }
+        } );
 
       });
 
@@ -153,10 +156,10 @@ describe('live-annotation', () => {
         const spy = sinon.spy();
         annotationService.on(EVENT_TRIGGERS.ANNOTATION_STROKE_DATA, spy);
 
-        await annotationService.processStrokeMessage({value:{encryptionKeyUrl: 'encryptionKeyUrl', content: 'content'}});
+        await annotationService.processStrokeMessage({request:{value:{encryptionKeyUrl: 'encryptionKeyUrl', content: 'content'}}});
 
         assert.calledOnceWithExactly(spy, {
-          payload: {encryptionKeyUrl: 'encryptionKeyUrl', content: 'decryptedContent'},
+          payload:{request:{value:{encryptionKeyUrl: 'encryptionKeyUrl', content: 'decryptedContent'}}} ,
         });
 
       });
@@ -341,20 +344,20 @@ describe('live-annotation', () => {
       });
 
       describe('change annotation options', () => {
-        it('makes change annotation options as expected', () => {
-          const options =  {
+        it('makes change annotation options as expected', async() => {
+          const options =  { annotationInfo:{
               version: '1',
               policy: 'AnnotationNotAllowed',
-          };
-          const meeting = {
-            meetingRequest : {
-              changeMeetingFloor: () => {
-              }
-            }
-          }
-          sinon.spy(meeting.meetingRequest, 'changeMeetingFloor');
-          annotationService.changeAnnotationOptions(options,meeting);
-          assert.calledOnceWithExactly(meeting.meetingRequest.changeMeetingFloor, options);
+          }};
+
+          const remoteShareUrl = 'remoteShareUrl';
+          const result = await annotationService.changeAnnotationOptions(remoteShareUrl,options);
+          assert.calledOnceWithExactly(webex.request, {
+              method: 'PATCH',
+              url: 'remoteShareUrl',
+              body: {annotationInfo: { version: '1', policy: 'AnnotationNotAllowed' }}
+          });
+          assert.equal(result, 'REQUEST_RETURN_VALUE')
         });
       });